+++ modules/spaint/src/ocv/OpenCVUtil.cpp
@@ -28,7 +28,7 @@ cv::Mat3b OpenCVUtil::make_rgb_image(const float *rgbData, int width, int height
   return result;
 }
 
-cv::Mat3b OpenCVUtil::pad_image(const cv::Mat3b& image, size_t paddingSize)
+cv::Mat3b OpenCVUtil::pad_image(const cv::Mat3b& image, int paddingSize)
 {
   cv::Mat3b paddedImage = cv::Mat3b::zeros(image.rows + paddingSize * 2, image.cols + paddingSize * 2);
   cv::Rect roi(paddingSize, paddingSize, image.cols, image.rows);
@@ -41,7 +41,7 @@ void OpenCVUtil::show_scaled_greyscale_figure(const std::string& windowName, con
   show_scaled_greyscale_figure(windowName, inputData, width, height, order, ScaleByFactor(scaleFactor));
 }
 
-cv::Mat3b OpenCVUtil::tile_image_patches(const std::vector<cv::Mat3b>& images, size_t tileCols, size_t tileRows, size_t patchWidth, size_t patchHeight, size_t paddingSize)
+cv::Mat3b OpenCVUtil::tile_image_patches(const std::vector<cv::Mat3b>& images, int tileCols, int tileRows, int patchWidth, int patchHeight, int paddingSize)
 {
   // Scale and pad the input images to make the tiles.
   std::vector<cv::Mat3b> tiles;
@@ -69,11 +69,11 @@ cv::Mat3b OpenCVUtil::tile_image_patches(const std::vector<cv::Mat3b>& images, s
   return tiledImage;
 }
 
-cv::Mat3b OpenCVUtil::tile_image_patches_bounded(const std::vector<cv::Mat3b>& images, size_t imageWidth, size_t imageHeight, size_t patchWidth, size_t patchHeight, size_t paddingSize)
+cv::Mat3b OpenCVUtil::tile_image_patches_bounded(const std::vector<cv::Mat3b>& images, int imageWidth, int imageHeight, int patchWidth, int patchHeight, int paddingSize)
 {
   // Compute the size of each tile.
-  size_t tileWidth = patchWidth + 2 * paddingSize;
-  size_t tileHeight = patchHeight + 2 * paddingSize;
+  int tileWidth = patchWidth + 2 * paddingSize;
+  int tileHeight = patchHeight + 2 * paddingSize;
 
   // If the image is too small to show even a single column or single row of tiles, throw.
   if(imageWidth < tileWidth || imageHeight < tileHeight)
@@ -82,8 +82,8 @@ cv::Mat3b OpenCVUtil::tile_image_patches_bounded(const std::vector<cv::Mat3b>& i
   }
 
   // Determine how many columns and rows of tiles we can fit into an output image of the specified size.
-  size_t tileCols = imageWidth / tileWidth;
-  size_t tileRows = imageHeight / tileHeight ;
+  int tileCols = imageWidth / tileWidth;
+  int tileRows = imageHeight / tileHeight ;
 
   // Tile the patches and return the output image.
   return tile_image_patches(images, tileCols, tileRows, patchHeight, patchWidth, paddingSize);
